refactor: updated OperationsMonitorDetails component
