+++ src/features/pre-auction/operations-monitor/operations-monitor-details.tsx
